feat/ui/category_data: list ads of specific category
