+++ src/services/frontend/src/components/footer/Footer.js
@@ -31,7 +31,7 @@ const Footer = () => {
                 <Typography variant="h4" align="center" mt={5}>
                     Follow Us
                 </Typography>
-                <Box align="center">Foll
+                <Box align="center">
                     <FacebookIcon sx={{ color: 'white', mr: 2 }} />
                     <TwitterIcon sx={{ color: 'white', mr: 2 }} />
                     <InstagramIcon sx={{ color: 'white', mr: 2 }} />
